factor out election screen
